保险产品/requirements.txt
# 保险产品RAG MVP系统依赖
# Python 3.8+

# PDF处理
pypdf2==3.0.1
pdfplumber==0.10.3

# LangChain生态
langchain==0.1.0
langchain-community==0.0.38
langchain-openai==0.1.0

# 向量存储
faiss-cpu==1.7.4
chromadb==0.4.22

# OpenAI
openai==1.12.0

# 数据处理
pandas==2.1.4
numpy==1.24.3

# 文本处理
tiktoken==0.5.2
python-dotenv==1.0.0

# 工具库
tqdm==4.66.1
pydantic==2.5.3
typing-extensions==4.9.0